\name{audiology}
\alias{audiology}
\docType{data}
\title{Audiology Dataset}
\description{This dataset contains the donations and charity information.}
\details{
The audiology dataset was from UCI machine learning repository. 
\dQuote{The dataset is in the details} 
\url{http://rexa.info/paper/1fb045668480ef3d72fef59f5f86bcb23a2ceffa}.
}
\usage{data(audiology)}
\source{http://rexa.info/paper/1fb045668480ef3d72fef59f5f86bcb23a2ceffa}
\format{data frame}
\value{A data frame of the audiology dataset}
\examples{
data("audiology")    #Load the audiology dataset
}
\keyword{audiology}
\keyword{uci}
