\name{charity}
\title{R Wrapper for Tetrad Library}
\description{
  The charity dataset.
}
\details{
The charity dataset was collected from a study of Cryder et al. (2013) \dQuote{The donor is in the details} \url{http://dx.doi.org/10.1016/j.obhdp.2012.08.002}, looking at what makes people donate more to charity. It is used as an example dataset in the \code{tetradR}. The intervention variable was TangibilityCondition and the other variables were recorded in a questionnaire using likert scales. 
}
\keyword{Decision making}
\keyword{Generosity}
\keyword{Charitable giving}
\keyword{Emotions}
\keyword{Impact}